Criacao das controllers 2

+++ src/main/java/iogianhenrique/repository/Alunos.java
@@ -3,5 +3,7 @@ package iogianhenrique.repository;
 import iogianhenrique.entity.Aluno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Alunos extends JpaRepository<Aluno,Integer> {
+import java.math.BigInteger;
+
+public interface Alunos extends JpaRepository<Aluno, BigInteger> {
 }
